+++ src/main/java/at/hannibal2/skyhanni/config/features/Chat.java
@@ -53,92 +53,13 @@ public class Chat {
     public boolean playerMessages = false;
 
     @Expose
-    @ConfigOption(name = "Message Format", desc = "")
-    @ConfigAccordionId(id = 1)
-    @ConfigEditorAccordion(id = 2)
-    public boolean messageFormat = false;
-
-    @Expose
-    @ConfigOption(name = "All Channel Prefix", desc = "Show the prefix for the all channel chat.")
-    @ConfigEditorBoolean
-    @ConfigAccordionId(id = 2)
-    public boolean allChannelPrefix = false;
-
-    @Expose
-    @ConfigOption(name = "Player Rank Hider", desc = "Hide player ranks in the chat.")
-    @ConfigEditorBoolean
-    @ConfigAccordionId(id = 2)
-    public boolean playerRankHider = false;
-
-    @Expose
-    @ConfigOption(name = "Player Colon Hider", desc = "Hide the colon after the player name in the chat.")
-    @ConfigEditorBoolean
-    @ConfigAccordionId(id = 2)
-    public boolean playerColonHider = false;
-
-    @Expose
-    @ConfigOption(
-            name = "SkyBlock Level Design",
-            desc = "Change the design of the Skyblock Level in the chat.\n" +
-                    "§cRequires 'SkyBlock Level in Chat' enabled in the SkyBlock Menu."
-    )
-    @ConfigEditorDropdown(
-            values = {"§8[§6123§8] §bname§f: msg",
-                    "§6§l123 §bname§f: §fmsg",
-                    "§bname §8[§6123§8]§f: msg",
-            "§cHide SkyBlock Level"}
-    )
-    @ConfigAccordionId(id = 2)
-    public int skyblockLevelDesign = 0;
-
-    @Expose
-    @ConfigOption(
-            name = "Elite Design",
-            desc = "Change the design of the Elite position in the chat."
-    )
-    @ConfigEditorDropdown(
-            values = {"§6[⌬499]",
-                    "§6§l⌬499",
-                    "§cHide Elite Position"}
-    )
-    @ConfigAccordionId(id = 2)
-    public int eliteFormat = 0;
-
-    @Expose
-    @ConfigOption(
-            name = "Channel Design",
-            desc = "Change the design of the Channel Prefix in the chat."
-    )
-    @ConfigEditorDropdown(
-            values = {"§2Guild >",
-                    "§2G>",
-                    "§8<§2G§8>",
-                    "§8[§2G§8]",
-                    "§8(§2G§8)"}
-    )
-    @ConfigAccordionId(id = 2)
-    public int channelDesign = 0;
-
-    @Expose
-    @ConfigOption(name = "Test All Chat", desc = "Test the All Chat message format locally (no message gets sent to hypixel)")
-    @ConfigEditorButton(runnableId = "testAllChat")
-    @ConfigAccordionId(id = 2)
-    public boolean testAllChat = false;
-
-    @Expose
-    @ConfigOption(name = "Test Guild Chat", desc = "Test the Guild Chat message format locally (no message gets sent to hypixel)")
-    @ConfigEditorButton(runnableId = "testGuildChat")
-    @ConfigAccordionId(id = 2)
-    public boolean testGuildChat = false;
-
-    @Expose
-    @ConfigOption(name = "NEU Profile Viewer", desc = "Click on a player name to open the Profile Viewer from NotEnoughUpdates")
+    @ConfigOption(name = "Player Rank Hider", desc = "Hide player ranks in all chat messages.")
     @ConfigEditorBoolean
     @ConfigAccordionId(id = 1)
-    public boolean neuProfileViewer = false;
+    public boolean playerRankHider = true;
 
     @Expose
-    @ConfigOption(name = "Chat Filter", desc = "Scan messages sent by players in all-chat for blacklisted words and greys out the message")
+    @ConfigOption(name = "Chat Filter", desc = "Scan messages sent by players for blacklisted words and grey out the message if any are found.")
     @ConfigEditorBoolean
     @ConfigAccordionId(id = 1)
     public boolean chatFilter = false;