+++ src/main/java/at/hannibal2/skyhanni/events/LorenzChatEvent.kt
@@ -2,4 +2,4 @@ package at.hannibal2.skyhanni.events
 
 import net.minecraft.util.IChatComponent
 
-class LorenzChatEvent(val message: String, val chatComponent: IChatComponent, var blockedReason: String = "") : LorenzEvent()
+class LorenzChatEvent(val message: String, var chatComponent: IChatComponent, var blockedReason: String = "") : LorenzEvent()