+++ src/main/java/at/hannibal2/skyhanni/features/chat/playerchat/PlayerChatFilter.kt
@@ -1,19 +1,27 @@
 package at.hannibal2.skyhanni.features.chat.playerchat
 
-import at.hannibal2.skyhanni.SkyHanniMod
-import at.hannibal2.skyhanni.events.PlayerSendChatEvent
 import at.hannibal2.skyhanni.events.RepositoryReloadEvent
 import at.hannibal2.skyhanni.utils.LorenzUtils
 import at.hannibal2.skyhanni.utils.MultiFilter
-import net.minecraft.util.ChatComponentText
-import net.minecraft.util.EnumChatFormatting
-import net.minecraftforge.fml.common.eventhandler.EventPriority
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
-import java.util.regex.Pattern
 
 class PlayerChatFilter {
 
-    private val filters = mutableMapOf<String, MultiFilter>()
+    companion object {
+
+        private val filters = mutableMapOf<String, MultiFilter>()
+
+        fun shouldChatFilter(original: String): Boolean {
+            val message = original.lowercase()
+            for (filter in filters) {
+                filter.value.matchResult(message)?.let {
+                    return true
+                }
+            }
+
+            return false
+        }
+    }
 
     @SubscribeEvent
     fun onRepoReload(event: RepositoryReloadEvent) {
@@ -41,49 +49,5 @@ class PlayerChatFilter {
             e.printStackTrace()
             LorenzUtils.error("error in RepositoryReloadEvent")
         }
-    }
-
-    @SubscribeEvent(priority = EventPriority.HIGH)
-    fun onPlayerChat(event: PlayerSendChatEvent) {
-        if (!SkyHanniMod.feature.chat.chatFilter) return
-        if (event.channel != PlayerMessageChannel.ALL) return
-
-        val message = event.message.lowercase()
-        for (filter in filters) {
-            filter.value.matchResult(message)?.let {
-                filter(event, it)
-                return
-            }
-        }
-    }
-
-    private fun filter(event: PlayerSendChatEvent, filter: String) {
-        val pattern = Pattern.compile("(.*)?$filter(.*)?", Pattern.CASE_INSENSITIVE)
-        val matcher = pattern.matcher(event.message)
-        matcher.matches()
-        val beginning = matcher.group(1)
-        val end = matcher.group(2)
-
-        event.chatComponents.clear()
-
-        val endSplit = end.split(" ")
-
-        for (word in beginning.split(" ")) {
-            if (word.isEmpty()) continue
-            event.chatComponents.add(coloredChat(word, EnumChatFormatting.GRAY))
-        }
-
-        event.chatComponents.add(coloredChat(filter.trim(), EnumChatFormatting.WHITE))
-
-        for (word in endSplit) {
-            if (word.isEmpty()) continue
-            event.chatComponents.add(coloredChat(word, EnumChatFormatting.GRAY))
-        }
-    }
-
-    private fun coloredChat(string: String, color: EnumChatFormatting): ChatComponentText {
-        val text = ChatComponentText(string)
-        text.chatStyle.color = color
-        return text
     }
 }